feat: fix linking
