+++ package.json
@@ -44,11 +44,11 @@
   },
   "devDependencies": {
     "@total-typescript/ts-reset": "^0.5.1",
+    "@trivago/prettier-plugin-sort-imports": "4.1.0",
     "@types/luxon": "^3.4.2",
     "@types/node": "20.12.7",
     "@types/react": "18.2.79",
     "@types/react-dom": "18.2.25",
-    "@trivago/prettier-plugin-sort-imports": "4.1.0",
     "eslint": "8.57.0",
     "eslint-config-next": "14.2.1",
     "prettier": "3.2.5",
